+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.4.0 IN PROGRESS
 
 * Server-side query sort (temporary, lacking i18n).
+* Correctly assign/unassign users to roles. Refs UIROLES-63, UIROLES-43.
 
 ## [1.3.0](https://github.com/folio-org/ui-authorization-roles/tree/v1.3.0) (2024-03-05)
 [Full Changelog](https://github.com/folio-org/ui-authorization-roles/compare/v1.2.0...v1.3.0)